+++ Editor/I18n/I18nBase.cs
@@ -105,6 +105,7 @@ namespace KRT.VRCQuestTools.I18n
         // Check for Update
         internal abstract string CheckLater { get; }
         internal abstract string GetUpdate { get; }
+        internal abstract string SeeChangelog { get; }
         internal abstract string SkipThisVersion { get; }
         internal abstract string NewVersionIsAvailable(string latestVersion);
         internal abstract string NewVersionHasBreakingChanges { get; }